# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangle API Documentation",
        "description": "API documentation for Tangle application. This API provides endpoints for managing collaborations, user profiles, reviews, and authentication.",
        "termsOfService": "https://tangle-api.example.com/terms",
        "contact": {
            "name": "Tangle API Support",
            "url": "https://tangle-api.example.com",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "operationId": "7ea792ba00ddd5f88aae2eb2acaf82c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}/follow": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Follow a user",
                "operationId": "bcef8495024830938e79edaa03e3fc9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow successful"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}/unfollow": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Unfollow a user",
                "operationId": "d1e10d83be66215110b04139edc1da24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unfollow successful"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user's basic profile information",
                "description": "Update basic profile information",
                "operationId": "56162de6c8a3b48291aee6cf5d6ee14a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "location",
                                    "bio"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "@JohnDoe"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "JohnDoe@gmail.com"
                                    },
                                    "tagline": {
                                        "type": "string",
                                        "example": "Developer Designer"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "London, UK"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "A brief introduction about yourself"
                                    },
                                    "profile_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[]": {
                                        "description": "Exactly 3 images required. Max 5MB each. Supported formats: jpeg, png, jpg",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        },
                                        "maxItems": 3,
                                        "minItems": 0
                                    },
                                    "social_media": {
                                        "properties": {
                                            "instagram": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "https://instagram.com/username",
                                                "nullable": true
                                            },
                                            "spotify": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "https://spotify.com/artist/username",
                                                "nullable": true
                                            },
                                            "tiktok": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "https://tiktok.com/@username",
                                                "nullable": true
                                            },
                                            "twitch": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "https://twitch.tv/username",
                                                "nullable": true
                                            },
                                            "youtube": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "https://youtube.com/c/username",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/completion": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user's profile completion status",
                "description": "Get profile completion status",
                "operationId": "f4493a8fc6bb42f04b9b74c755f2b074",
                "responses": {
                    "200": {
                        "description": "Profile completion status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "completion_status": {
                                            "properties": {
                                                "basic_info": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "portfolio": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "social_media": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "completion_percentage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 66.67
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}